oculto-social/pyproject.toml
[project]
name = "oculto-social"
version = "0.1.0"
description = ""
authors = [
    {name = "redscips",email = "redscipher@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "markdown (>=3.8,<4.0)",
    "django-filter (>=25.1,<26.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "django-debug-toolbar (>=5.2.0,<6.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
]

[tool.poetry]
packages = [{include = "ocultosocial"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"